+++ cypress/e2e/maputnik-driver.ts
@@ -94,8 +94,8 @@ export class MaputnikDriver {
   public when = {
     ...this.helper.when,
     modal: this.modalDriver.when,
-    within: (selector: string, fn: () => void) => {
-      this.helper.when.within(fn, selector);
+    doWithin: (selector: string, fn: () => void) => {
+      this.helper.when.doWithin(fn, selector);
     },
     tab: () => this.helper.get.element("body").tab(),
     waitForExampleFileResponse: () => {
@@ -110,25 +110,25 @@ export class MaputnikDriver {
       styleProperties: "geojson" | "raster" | "both" | "layer" | "",
       zoom?: number
     ) => {
-      let url = "?debug";
+      const url = new URL(baseUrl);
       switch (styleProperties) {
       case "geojson":
-        url += `&style=${baseUrl}geojson-style.json`;
+        url.searchParams.set("style", baseUrl + "geojson-style.json");
         break;
       case "raster":
-        url += `&style=${baseUrl}raster-style.json`;
+        url.searchParams.set("style", baseUrl + "raster-style.json");
         break;
       case "both":
-        url += `&style=${baseUrl}geojson-raster-style.json`;
+        url.searchParams.set("style", baseUrl + "geojson-raster-style.json");
         break;
       case "layer":
-        url += `&style=${baseUrl}/example-layer-style.json`;
+        url.searchParams.set("style", baseUrl + "example-layer-style.json");
         break;
       }
       if (zoom) {
-        url += `#${zoom}/41.3805/2.1635`;
+        url.hash = `${zoom}/41.3805/2.1635`;
       }
-      this.helper.when.visit(baseUrl + url);
+      this.helper.when.visit(url.toString());
       if (styleProperties) {
         this.helper.when.acceptConfirm();
       }
@@ -145,7 +145,7 @@ export class MaputnikDriver {
     },
 
     selectWithin: (selector: string, value: string) => {
-      this.when.within(selector, () => {
+      this.when.doWithin(selector, () => {
         this.helper.get.element("select").select(value);
       });
     },