+++ src/libs/store/apistore.ts
@@ -1,46 +1,38 @@
-import style from './style.js'
+import style from '../style'
 import {format} from '@maplibre/maplibre-gl-style-spec'
-import type {StyleSpecification} from 'maplibre-gl'
 import ReconnectingWebSocket from 'reconnecting-websocket'
+import type {IStyleStore, OnStyleChangedCallback, StyleSpecificationWithId} from '../definitions'
 
 export type ApiStyleStoreOptions = {
-  port: string | null
-  host: string | null
-  onLocalStyleChange?: (style: any) => void
+  onLocalStyleChange?: OnStyleChangedCallback
 }
 
-export class ApiStyleStore {
+export class ApiStyleStore implements IStyleStore {
 
   localUrl: string;
   websocketUrl: string;
   latestStyleId: string | undefined = undefined;
-  onLocalStyleChange: (style: any) => void;
+  onLocalStyleChange: OnStyleChangedCallback;
 
   constructor(opts: ApiStyleStoreOptions) {
     this.onLocalStyleChange = opts.onLocalStyleChange || (() => {})
-    const port = opts.port || '8000'
-    const host = opts.host || 'localhost'
+    const port = window.location.port
+    const host = 'localhost'
     this.localUrl = `http://${host}:${port}`
     this.websocketUrl = `ws://${host}:${port}/ws`
     this.init = this.init.bind(this)
   }
 
-  init(cb: (...args: any[]) => void) {
-    fetch(this.localUrl + '/styles', {
-      mode: 'cors',
-    })
-      .then((response) =>  {
-        return response.json();
-      })
-      .then((body) => {
-        const styleIds = body;
-        this.latestStyleId = styleIds[0]
-        this.notifyLocalChanges()
-        cb(null)
-      })
-      .catch(() => {
-        cb(new Error('Can not connect to style API'))
-      })
+  async init(): Promise<void> {
+    try {
+      const response = await fetch(this.localUrl + '/styles', {mode: 'cors'});
+      const body = await response.json();
+      const styleIds = body;
+      this.latestStyleId = styleIds[0]
+      this.notifyLocalChanges();
+    } catch {
+      throw new Error('Can not connect to style API');
+    }
   }
 
   notifyLocalChanges() {
@@ -59,24 +51,20 @@ export class ApiStyleStore {
     }
   }
 
-  latestStyle(cb: (...args: any[]) => void) {
+  async getLatestStyle(): Promise<StyleSpecificationWithId> {
     if(this.latestStyleId) {
-      fetch(this.localUrl + '/styles/' + this.latestStyleId, {
+      const response = await fetch(this.localUrl + '/styles/' + this.latestStyleId, {
         mode: 'cors',
-      })
-        .then(function(response) {
-          return response.json();
-        })
-        .then(function(body) {
-          cb(style.ensureStyleValidity(body))
-        })
+      });
+      const body = await response.json();
+      return style.ensureStyleValidity(body);
     } else {
       throw new Error('No latest style available. You need to init the api backend first.')
     }
   }
 
   // Save current style replacing previous version
-  save(mapStyle: StyleSpecification & { id: string }) {
+  save(mapStyle: StyleSpecificationWithId) {
     const styleJSON = format(
       style.stripAccessTokens(
         style.replaceAccessTokens(mapStyle)